requirements.txt
aioredis == 2.0.0
asyncpg == 0.28.0
bcrypt == 4.0.1
fastapi == 0.100.1
Jinja2 == 3.1.2
loguru == 0.7.0
passlib == 1.7.4
psycopg2 == 2.9.6
pydantic-settings == 2.0.2
python-decouple == 3.8
python-jose == 3.3.0
python-multipart == 0.0.6
sqlalchemy == 2.0.19
uvicorn == 0.23.2
validators == 0.20.0
python-ulid == 1.1.0
#websockets == 10.3
#wsproto == 1.1.0
